refactor: 모든 엔티티 및 설정 패키지 이름 변경
`entity`와 `config` 패키지를 `org.yechan.entity`, `org.yechan.config`로 변경하여 네이밍 컨벤션 통일 및 프로젝트 구조 개선.
