--- domain-core/src/main/java/entity/BaseEntity.java
@@ -1,26 +1,0 @@
-package entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-@Getter
-@NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime created_at;
-
-    @LastModifiedDate
-    @Column(nullable = false)
-    private LocalDateTime modified_at;
-}
