--- domain-order/src/main/java/entity/OrderItem.java
@@ -1,42 +1,0 @@
-package entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "order_items")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OrderItem extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "order_item_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id")
-    private Order order;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "seat_id")
-    private Seat seat;
-
-    @Column(name = "seat_name", nullable = false)//TODO 2025 05 22 20:58:27 : custom value and converter
-    private String seatName; // 좌석 이름 (예: A열 12번)
-
-    @Column(name = "seat_grade", nullable = false)
-    private String seatGrade; // 좌석 등급 (VIP, R, S 등)
-
-    @Column(name = "price", nullable = false)
-    private Integer price; // 좌석 가격
-}
