--- domain-payment/src/main/java/entity/PaymentStatus.java
@@ -1,8 +1,0 @@
-package entity;
-
-public enum PaymentStatus {
-    PENDING,
-    APPROVED,
-    FAILED,
-    CANCELED;
-}
