+++ domain-payment/src/main/java/org/yechan/entity/PaymentStatus.java
@@ -1,0 +1,8 @@
+package org.yechan.entity;
+
+public enum PaymentStatus {
+    PENDING,
+    APPROVED,
+    FAILED,
+    CANCELED;
+}
