+++ domain-user/src/main/java/org/yechan/repository/JpaUserRepository.java
@@ -1,0 +1,11 @@
+package org.yechan.repository;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import org.yechan.entity.User;
+
+@Repository
+public interface JpaUserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String username);
+}
