+++ domain-user/src/main/java/org/yechan/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package org.yechan.repository;
+
+public interface UserRepository extends JpaUserRepository {
+}
