--- domain-user/src/main/java/repository/JpaUserRepository.java
@@ -1,11 +1,0 @@
-package repository;
-
-import entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface JpaUserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String username);
-}
