--- domain-user/src/main/java/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public interface UserRepository extends JpaUserRepository {
-}
