--- internal/internal-security/src/main/java/config/SecurityFilterRegistrar.java
@@ -1,8 +1,0 @@
-package config;
-
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-
-@FunctionalInterface
-public interface SecurityFilterRegistrar {
-    void register(HttpSecurity http);
-}
