+++ internal/internal-security/src/main/java/org/yechan/config/SecurityFilterRegistrar.java
@@ -1,0 +1,8 @@
+package org.yechan.config;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+
+@FunctionalInterface
+public interface SecurityFilterRegistrar {
+    void register(HttpSecurity http);
+}
