+++ internal/internal-security/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  config.SecurityAutoConfiguration
+  org.yechan.config.SecurityAutoConfiguration
