+++ internal/internal-security/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-config.SecurityAutoConfiguration
+org.yechan.config.SecurityAutoConfiguration
